requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
mailjet-rest==1.3.4
psycopg2==2.9.10
python-dotenv==1.1.0